Fix stage percentages

+++ src/Angor/Avalonia/Angor.Contexts.Funding/Projects/Domain/Stage.cs
@@ -4,5 +4,5 @@ public class Stage
 {
     public DateTime ReleaseDate { get; set; }
     public int Index { get; set; }
-    public decimal RatioOfTotal { get; set; }
+    public double RatioOfTotal { get; set; }
 }
